pyproject.toml
[project]
name = "the-nomad-data-stack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "pyyaml>=6.0.2",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "akshare>=1.17.96",
    "prefect>=3.0.0",
    "boto3>=1.34.0",
]

[dependency-groups]
dev = [
    "apache-airflow>=3.1.5",
    "apache-airflow-providers-amazon>=9.18.1",
    "apache-airflow-providers-standard>=1.0.0",
    "freezegun>=1.5.5",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
